requirements.txt
tensorflow==2.19.0
numpy==2.1.3
pillow==11.2.1
streamlit==1.35.0
matplotlib==3.10.3
scikit-learn==1.7.0
pandas==2.3.0
protobuf==5.29.5
h5py==3.14.0
